# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SILICOFELLER Quantum Studio API",
        "description": "AI-augmented quantum hardware EDA platform — V2. DesignGraph → Constraints → Placement → Routing → DRC → Qiskit Metal → Tapeout.",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/simulations": {
            "get": {
                "tags": [
                    "simulations"
                ],
                "summary": "List Simulations",
                "operationId": "list_simulations_api_simulations_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "additionalProperties": true,
                                        "type": "object"
                                    },
                                    "type": "array",
                                    "title": "Response List Simulations Api Simulations Get"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Create Simulation",
                "operationId": "create_simulation_api_simulations_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SimulationCreate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Create Simulation Api Simulations Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/api/simulations/{sim_id}/run": {
            "post": {
                "tags": [
                    "simulations"
                ],
                "summary": "Run Simulation",
                "description": "Run a simulation analytically (physics engine).\nIn production this would dispatch to Palace/scqubits workers.",
                "operationId": "run_simulation_api_simulations__sim_id__run_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "sim_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Sim Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Run Simulation Api Simulations  Sim Id  Run Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/simulations/{sim_id}": {
            "get": {
                "tags": [
                    "simulations"
                ],
                "summary": "Get Simulation",
                "operationId": "get_simulation_api_simulations__sim_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "sim_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Sim Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Simulation Api Simulations  Sim Id  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/design/route": {
            "post": {
                "tags": [
                    "design-v2"
                ],
                "summary": "Route Placement",
                "description": "Run routing on a placement dict (from /api/generate/placement).\nReturns coupler, resonator, and feedline route segments.",
                "operationId": "route_placement_api_design_route_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RouteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": true,
                                    "type": "object",
                                    "title": "Response Route Placement Api Design Route Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RouteRequest": {
                "properties": {
                    "placement": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Placement"
                    },
                    "constraints": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Constraints"
                    }
                },
                "type": "object",
                "required": [
                    "placement"
                ],
                "title": "RouteRequest"
            },
...
            "SimulationCreate": {
                "properties": {
                    "project_id": {
                        "type": "string",
                        "title": "Project Id"
                    },
                    "solver": {
                        "type": "string",
                        "title": "Solver",
                        "default": "eigenmode"
                    },
                    "config": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Config",
                        "default": {}
                    }
                },
                "type": "object",
                "required": [
                    "project_id"
                ],
                "title": "SimulationCreate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/auth/token"
                    }
                }
            }
        }
    }
}